Add check for yarn version in createLessonPdfs, and require at least version 1.3.2

+++ createLessonPdfs.js
@@ -1,19 +1,18 @@
 /* eslint-env node */
 
-const { spawn } = require('child_process');
+const { spawn, execSync } = require('child_process');
 const nodeCleanup = require('node-cleanup');
 const puppeteer = require('puppeteer');
 const path = require('path');
 const fse = require('fs-extra');
 const {buildDir, publicPath} = require('./buildconstants');
-
 const {lessonPaths} = require('./pathlists');
 
 const urlBase = 'http://127.0.0.1:8080' + publicPath;
-
 let localWebServer = null;
+
 const cleanup = () => {
-  if (!localWebServer.killed) {
+  if (localWebServer && !localWebServer.killed) {
     console.log('Killing localWebServer');
     const killSignal = 'SIGTERM'; // 'SIGINT';
     const killSuccess = localWebServer.kill(killSignal);
@@ -23,10 +22,6 @@ const cleanup = () => {
     localWebServer.removeAllListeners();
   }
 };
-nodeCleanup(function (exitCode, signal) {
-  console.log('Exiting node script... (exitCode:' + exitCode + ', signal:' + signal + ')');
-  cleanup();
-});
 
 const convertUrl = async (browser, lesson) => {
   try {
@@ -71,10 +66,6 @@ const doConvert = () => {
 
 };
 
-localWebServer = spawn('yarn serve', {
-  shell: true
-});
-
 const checkStarted = (data) => {
   const str = String(data);
   console.log('localWebServer:', str);
@@ -82,6 +73,31 @@ const checkStarted = (data) => {
     doConvert(localWebServer);
   }
 };
+
+const checkYarnVersion = () => {
+  const version = execSync('yarn --version', {shell: true}).toString().trim();
+  const lowestVersion = '1.3.2';
+  const [major = 0, minor = 0, patch = 0] = version.split('.');
+  const [lowestMajor = 0, lowestMinor = 0, lowestPatch = 0] = lowestVersion.split('.');
+  const tooLow = () => {
+    console.log('ERROR: The version of yarn (' + version + ') is too low. Must be >= ' + lowestVersion);
+    process.exit(1);
+  };
+  if (major < lowestMajor) { tooLow(); }
+  if (major === lowestMajor && minor < lowestMinor) { tooLow(); }
+  if (major === lowestMajor && minor === lowestMinor && patch < lowestPatch) { tooLow(); }
+};
+
+checkYarnVersion();
+
+nodeCleanup(function (exitCode, signal) {
+  console.log('Exiting node script... (exitCode:' + exitCode + ', signal:' + signal + ')');
+  cleanup();
+});
+
+localWebServer = spawn('yarn serve', {
+  shell: true
+});
 
 localWebServer.stdout.on('data', checkStarted);
 localWebServer.stderr.on('data', checkStarted);
